pyproject.toml
[tool.poetry]
name = "documentation-helper"
version = "0.1.0"
description = ""
authors = ["chmiller <chris.wayne.miller@gmail.com>"]
readme = "README.md"
packages = [
     {include = "documentation_helper"},
     {include = "backend"},
     {include = "utils"}
]

[tool.poetry.dependencies]
python = "^3.11"
langchain = "^0.0.180"
beautifulsoup4 = "^4.12.2"
black = "^23.3.0"
tiktoken = "^0.4.0"
openai = "^0.27.7"
pinecone-client = "^2.2.1"
unstructured = "^0.6.9"
nltk = "^3.8.1"
fastapi = "^0.95.2"
Jinja2 = "^3.1.2"
uvicorn = "^0.22.0"
streamlit = "^1.22.0"
streamlit-chat = "^0.0.2.2"
tqdm = "^4.65.0"
python-dotenv = "^1.0.0"
toml = "^0.10.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
